pyproject.toml
[project]
name = "lawchat"
version = "1.0.0"
description = "법률 상담을 위한 AI 어시스턴트"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "llama-index>=0.9.0",
    "llama-index-llms-openai",
    "llama-index-embeddings-openai",
    "llama-index-vector-stores-chroma",
    "sqlalchemy>=2.0.0",
    "mysql-connector-python",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "httpx-oauth",
    "PyJWT",
    "nest-asyncio"
]
